+++ src/elecraft/profile.rs
@@ -57,6 +57,8 @@ pub struct ElecraftProfile {
     pub attenuator_max: Option<u8>,
     pub supports_noise_blanker: bool,
     pub supports_agc: bool,
+    pub filter_max_hz: Option<u16>,
+    pub filter_command: &'static str,
     pub af_gain_max: Option<u16>,
     pub rf_gain_max: Option<u16>,
     pub squelch_max: u16,
@@ -107,6 +109,7 @@ impl ElecraftProfile {
             ControlId::Attenuator => self.attenuator_max.is_some(),
             ControlId::NoiseBlanker => self.supports_noise_blanker,
             ControlId::Agc => self.supports_agc,
+            ControlId::Filter => self.filter_max_hz.is_some(),
             ControlId::Squelch
             | ControlId::RfPower
             | ControlId::Vfo